refactor(WikiList): 위키리스트 리팩토링
