+++ src/_components/MaeilWiki/Home/home.css.ts
@@ -1,5 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
+// HeroSection
 export const heroSection = style({
   background: `rgba(0, 168, 107, 0.08)`,
   borderRadius: '1rem',
@@ -12,4 +13,10 @@ export const title = style({
   color: '#00a86b',
   fontSize: '2.8rem',
   fontWeight: 700,
+});
+
+export const subtitle = style({
+  color: '#545454',
+  fontSize: '1.5rem',
+  marginTop: '1.3rem',
 });
